+++ src/components/GrepEditor/handlers/key.ts
@@ -6,11 +6,27 @@ import {
 } from 'draft-js';
 import { Style } from '../buttons';
 
+export type CustomDraftCommand = DraftEditorCommand | 'shift-split-block';
+
+export const customKeyHandler = (
+  setEditorState: (state: EditorState) => void,
+) => (
+  command: CustomDraftCommand,
+  editorState: EditorState,
+): DraftHandleValue => {
+  if (command === 'shift-split-block') {
+    const newState = RichUtils.insertSoftNewline(editorState);
+    setEditorState(newState);
+    return 'handled';
+  }
+  return 'not-handled';
+};
+
 export const keyHandler = (
   setEditorState: (state: EditorState) => void,
   allowedStyles?: Array<Style>,
 ) => (
-  command: DraftEditorCommand,
+  command: CustomDraftCommand,
   editorState: EditorState,
 ): DraftHandleValue => {
   if (!allowedStyles || allowedStyles.some((s) => s === command)) {
@@ -20,7 +36,7 @@ export const keyHandler = (
       return 'handled';
     }
   }
-  return 'not-handled';
+  return customKeyHandler(setEditorState)(command, editorState);
 };
 
 export default keyHandler;